Add checks for user defined indexes.

+++ contrib/pg_upgrade/gp/check_greenplum_internal.c
@@ -1,3 +1,11 @@
+/*
+ *
+ *  check_greenplum_internal.c
+ *
+ *	Copyright (c) 2019-Present Pivotal Software, Inc
+ *
+ */
+
 #include "pg_upgrade.h"
 
 
@@ -18,9 +26,12 @@ perform_check(check_function check, ClusterInfo *cluster)
 
 
 void
-perform_greenplum_checks(check_function gp_checks_list[], int gp_checks_list_length, ClusterInfo *cluster)
+perform_greenplum_checks(
+	check_function gp_checks_list[],
+	int gp_checks_list_length,
+	ClusterInfo *cluster)
 {
-	size_t i;
+	int i;
 
 	for (i = 0; i < gp_checks_list_length; i++)
 	{
